pyproject.toml
[tool.poetry]
name = "wallet-open-test"
version = "0.1.0"
description = "Benchmarking wallet create and open times"
authors = ["Daniel Bluhm <dbluhm@pm.me>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
python3-indy = "^1.16.0"
aries-askar = "^0.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
